Lots of core changes (Long has been substitued in stead of Date in a lot of places; it shouldn't have been necessary but it seemed to at the time.)
Note that to open the editor you double click on a company in the list.

Add a company via the file menu, using the token.

Before the model can be run from the "editor", you must generate
historical data, by right clicking on a company and clicking build
historical. Enter dates you want, and bam.